requirements.txt
anyio==3.7.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
deezer-python==7.0.0
deezloader==2020.11.26
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.12.0
httpcore==0.13.7
httpx==1.0.0b0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
numpy==2.1.0
opencv-python==4.10.0.84
pycryptodome==3.20.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.14
redis==5.1.0b7
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
spotipy==2.24.0
stagger==1.0.0
tqdm==4.66.5
urllib3==2.2.2
Werkzeug==3.0.4
